faas/requirements.txt
Flask==3.1.0
gunicorn==23.0.0 
cloud-sql-python-connector==1.17.0
PyMySQL==1.1.1
google-cloud-storage==3.1.0
functions-framework==3.8.2
